Fix asynchronous event processing thread bug
An attempt to overwrite data while the thread was in its last phase would not trigger a restart of the event processing thread.

Fixed by adding an extra check at the end of the synchronous exit and also removing the redundant safeContinue flag, as a nil check on the tempData also suffices.